... adding documentation and changing the name to do so.
"...Although, there are a few topological errors (overlaps specifically) at each individual admin levels yet these overlaps are so small that are bearable for the tool so we may ignore them

Second problem was that cell size of raster is too big especially for STAT_LEVL_=3. The tool only calculates statistics for feature with highest number of OID if multiple features fall into a single raster cell.

Solution:- To overcome this issue, input raster needs to be resampled to a much higher resolution. Tiff files produced from 1st step/code has cell size of 0.9375 x 0.94......

at this resolution the tool calculates statistics for only 920 features out of 1920

resampling this cell size to 0.01 x 0.01 makes the tool to take around 15 minutes but calculates stats for each feature/polygon

hit and trial method lead to the conclusion that 0.04 is the optimum/maximum allowable cell size for which the tool can calculate stats for each feature

let me summarize my findings below.
Left number shows cell size while middle number shows the time it takes to complete the process for one day and right number shows the number of features it calculates stats for....
0.09 size 15 secs 1894 features
0.07 size 15 secs 1908 features
0.05 size 18.28 secs 1916 features
0.03 size 38.02 secs 1920 features
0.04 size 22 secs 1920 features