April updates - refactored for memory issues
Refined further to deal with March running errors, and errors for combining CSVs.

Instead of writing each file in memory and then combining them, the updated script now reads only 1 file into memory at a time, appends it to your defined output file and repeats.

This way possibility of getting a memory error is ruled out

The major drawback of this approach, though, is; as the defined output file keeps getting built up, it takes more and more time to append each later csv to it.

Ex: it took around 9 hours to merge only 288 files on my end.