June refactoring to overcome ArcPy/Map memory errors - refactoring
Here code was further refactored. The part 2. of the code/toolbox was broken into multiple files in order to handle memory/processing better. Below are the notes:

"1. ....tested the code on different machines in my office and strangely enough, it exhibited different behaviours. On some machines, it ran successfully. On some, it have the same error as yours and on some other machines it made arcmap totally crash. The behaviour is strange as I am clearing memory in every way I could think of.

2. ... then refactored the code such that I created smaller functions for each step and then just called the functions in the main body of code. I have put this code on testing on a machine but it's taking too long to process. It ran for 30 hours straight and processed only 279 files. That's another surprise. Theoretically, it shouldn't have made any difference in the processing time but it took so long.

3. I then tried another approach. I splitted the code no. 2 into 3 codes. I have seperated the part that was generating error that is the module / native tool "extract values to points". You may now use code 2a to calculate prate stats and get the output into shapefiles for each day. Use script 2b to convert these shapefiles into csv after necessary modifications. And use 2c to calculate stats for polygon shapefile exactly as before.

 210619_refactored.zip is the associated file.