Update of the 2a SHP file code for starting at folders
Last version used to prepare whole list of files present in the input folder (whether lying at its root or in subdirectories) and then process each file, one by one from the list.

If the folder contained 100 years of data, that means the list would contain 36,500 records corresponding to each file (with full path). Although, this much number of records shouldn't be a problem to handle theoretically, yet that might be the issue causing crashing of ArcGIS midway.

The attached script avoids it by creating a list of 1 year only, processes each file of 1 year and then creates a new list for next year, processes it and so on.